refactor: modularize `addMessageToDiv`
